Port PDF viewer from Evince to Papers
Evince has no GTK4 port, and upstream has moved to Papers as its
replacement. Drop evinceadapter.py, add papersadapter.py using
PapersDocument/PapersView 4.0 instead of EvinceDocument/EvinceView
3.0.

Split view init into PapersView.DocumentModel + PapersView.View,
since Papers separates document state from the view widget in a way
Evince didn't - EvinceView.View held both together.

Rewrite touch/gesture handling with Gtk.GestureClick/Gtk.GestureDrag,
since GTK4 removed the event signal and Gdk.EventMask.TOUCH_MASK this
depended on. Work around a Papers pointer-grab bug where the drag
gesture doesn't release properly - see the comment in
setup_gestures() for details.

Create the SearchContext before the document loads, not after -
Papers wires per-page highlighting into the child PpsViewPage widgets
during load, and creating it later leaves search highlighting broken.

load_document() now converts paths to URIs explicitly, since Papers
is stricter about accepting only URIs than Evince was.

Known issue: TOC is disabled. Papers moved the outline from
GtkTreeModel to GListModel, and has_document_links() segfaults on
GListModel input. Needs a full rewrite of the outline parser - not
attempting that in this commit.
